requirements.txt
streamlit
deepface
opencv-python-headless
tensorflow
pillow
